Formatting, whitespace, move waiting dec out of util
